+++ ydb/core/kqp/ut/batch_operations/kqp_batch_update_ut.cpp
@@ -1065,7 +1065,7 @@ Y_UNIT_TEST_SUITE(KqpBatchUpdate) {
         {
             auto result = session.ExecuteQuery(R"(
                 CREATE TABLE fulltext_plain_idx (
-                    k Int32 NOT NULL,
+                    k Uint64 NOT NULL,
                     v1 String,
                     v2 String,
                     v3 String,
@@ -1077,7 +1077,7 @@ Y_UNIT_TEST_SUITE(KqpBatchUpdate) {
                 );
 
                 CREATE TABLE fulltext_relevance_idx (
-                    k Int32 NOT NULL,
+                    k Uint64 NOT NULL,
                     v1 String,
                     v2 String,
                     v3 String,
